% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_CODE.R
\name{cal_aucc}
\alias{cal_aucc}
\title{Calculate AUCC value for data}
\usage{
cal_aucc(lista_same_filter)
}
\arguments{
\item{lista_same_filter}{list, top 500 DE genes detected by every DE method, calculated within the same filtering method group.}
}
\value{
AUCC (Area Under Consistency Curve) results.
}
\description{
Calculate AUCC value for data
}
