% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_CODE.R
\name{cal_cdo}
\alias{cal_cdo}
\title{Calculate CDO values for data}
\usage{
cal_cdo(listall)
}
\arguments{
\item{listall}{list, top 500 DE genes detected by every combination of methods.}
}
\value{
CDO (Consistent DE gene Order) result for every method
}
\description{
Calculate CDO values for data
}
