% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CODE_plot.R
\name{CODE_plot}
\alias{CODE_plot}
\title{CODE plot function}
\usage{
CODE_plot(list)
}
\arguments{
\item{CODE_list}{results of run_CODE}
}
\value{
Heatmap of AUCC

Heatmap of CDO

Heatmap of Metrics (AUCC+CDO)
}
\description{
Plot the results of run_CODE
}
\author{
Jiawei Zou
}
