% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_sample}
\alias{data_sample}
\title{A scRNA-seq dataset containg activated and naive CD4+ T cells}
\format{
A conut matrix 13045 genes x 462 cells:
\describe{
\item{row}{genes}
\item{col}{cells}
...
}
}
\source{
\url{https://science.sciencemag.org/content/355/6332/1433}
}
\usage{
data_sample
}
\description{
A scRNA-seq dataset containg activated and naive CD4+ T cells. 13045 genes x 462 cells.
}
\keyword{scRNA-seq dataset}
