% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scCODE-data.R
\docType{data}
\name{data_sccode}
\alias{data_sccode}
\title{A scRNA-seq dataset containg activated and naive CD4+ T cells}
\format{
A conut matrix 13045 genes x 462 cells:
\describe{
\item{row}{gene name}
\item{column}{cell name}
...
}
}
\source{
\url{https://www.science.org/doi/full/10.1126/science.aah4115}
}
\usage{
data_sccode
}
\description{
A scRNA-seq dataset containg activated and naive CD4+ T cells. 13045 genes x 462 cells.
}
\keyword{datasets}
