% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter_functions.R
\name{filter}
\alias{filter}
\alias{scCODE.filter_OGFSC}
\alias{scCODE.filter_conquer}
\alias{scCODE.filter_scmap}
\title{Gene filtering methods used in CODE}
\usage{
scCODE.filter_OGFSC(datarow)

scCODE.filter_conquer(datarow)

scCODE.filter_scmap(datarow)
}
\arguments{
\item{datarow}{matrix, data matrix after filtering.}
}
\value{
idx, the idx of remained genes.
}
\description{
Gene filtering methods used in CODE (Consensus Optimization of Differentially Expressed gene detection).
}
\examples{
set.seed(123)
datarow=data_sample
scCODE.filter_OGFSC(datarow)
}
\author{
Jiawei Zou
}
