% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{group_sample}
\alias{group_sample}
\title{Group information of cells}
\format{
A factor length 462
\describe{
\item{1}{activated CD4T+ cells}
\item{2}{naive CD4T+ cells}
...
}
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
group_sample
}
\description{
A vector of 1 and 2, the group information of the cells. Here, 1 for activated CD4T+ cells, 2 for naive CD4T+ cells.
}
\keyword{datasets}
