% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_CODE.R
\name{run_CODE}
\alias{run_CODE}
\title{Consensus Optimization of Differentially Expressed gene detection in single-cell RNA-seq data}
\usage{
run_CODE(datarow, group, light = TRUE, outdir, nts = 5)
}
\arguments{
\item{datarow}{matrix, data matrix after filtering.}

\item{group}{vector, cell type information (eg: 1 for cell type 1, 2 for cell type 2)}

\item{light}{True or False, run CODE in a light version (apply on part of methods which averagely perform well), default as False (runing with all methods).}

\item{outdir}{Path to save DE gene identification results.}

\item{nts}{number of top-ranked strategies for consensus optimization, 5 =< nts <= 10.}
}
\value{
AUCC, matrix results of combinations of methods.

CDO, matrix results of combinations of methods.

Optimal solution, optimal combination of methods for the data.

Optimal results, optimal DE results detected by the optmal solution. ## DE gene name, p.adj, foldchange (cell type 1 vs celltype2, appeared firstly in group would be celltype1).
}
\description{
CODE presents personalized evaluation of different combination of gene filtering methods and DE methods for individual scRNA-seq data,
and provides with personalized optimal DE result.
}
\examples{
set.seed(123)
run_CODE(data_sample,group_sample,light=TRUE,outdir='./',nts=5)
}
\author{
Jiawei Zou
}
