% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scCODE.R
\name{scCODE}
\alias{scCODE}
\title{Consensus Optimization of Differentially Expressed gene detection in single-cell RNA-seq data}
\usage{
scCODE(data, group, light = TRUE, top_ranked = 5)
}
\arguments{
\item{data}{matrix, data matrix after filtering.}

\item{group}{vector, cell type information (eg: 1 for cell type 1, 2 for cell type 2)}

\item{light}{True or False, run CODE in a light version (apply on part of methods which averagely perform well), default as False (runing with all methods).}

\item{top_ranked}{number of top-ranked strategies for consensus optimization, 5 <= top_ranked <= 10.}
}
\value{
DE results, re-ranked DE results detected by the top-ranked solutions.

Z-score-metrics, Z-scores of the CDO and AUCC.
}
\description{
scCODE makes data-specific evaluation of different combination of gene filtering methods and DE methods scRNA-seq data,
and provides with data-specific optimal DE result.
}
\examples{

set.seed(123)
test=scCODE(data_sccode,group_sccode,light=TRUE,top_ranked=5)
head(test$DE_results)

}
\author{
Jiawei Zou
}
