% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/New_Metrics.R
\name{scCODE_cdo}
\alias{scCODE_cdo}
\title{Calculate CDO values for data}
\usage{
scCODE_cdo(listall, list_order)
}
\arguments{
\item{listall}{list, top 500 DE genes detected by every combination of methods.}

\item{list_order, }{list, ALL gene rank for each analysis strategy.}
}
\value{
CDO (Consistent DE gene Order) result for every method
}
\description{
Calculate CDO values for data
}
