% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scCODE_plot.R
\name{scCODE_plot}
\alias{scCODE_plot}
\title{CODE plot function}
\usage{
scCODE_plot(list)
}
\arguments{
\item{list}{results of run_CODE}
}
\value{
Heatmap of AUCC

Heatmap of CDO

Heatmap of Metrics (AUCC+CDO)
}
\description{
Plot the results of run_CODE
}
\author{
Jiawei Zou
}
