% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_functions.R
\name{temp}
\alias{temp}
\alias{CODE.BPSC}
\alias{CODE.t_test}
\alias{CODE.DESeq2}
\alias{CODE.MAST}
\alias{CODE.wilcox_test}
\alias{CODE.DEsingle}
\alias{CODE.edgeR}
\alias{CODE.samr}
\alias{CODE.limma}
\alias{CODE.scDD}
\title{DE methods used in CODE}
\usage{
CODE.BPSC(datarow, group)

CODE.t_test(datarow, group)

CODE.DESeq2(datarow, group)

CODE.MAST(datarow, group)

CODE.wilcox_test(datarow, group)

CODE.DEsingle(datarow, group)

CODE.edgeR(datarow, group)

CODE.samr(datarow, group)

CODE.limma(datarow, group)

CODE.scDD(datarow, group)
}
\arguments{
\item{datarow}{matrix, data matrix after filtering.}

\item{group}{factor, cell group information for data}
}
\value{
results, the p value for each gene.
}
\description{
DE methods used to detect DE genes for scRNA-seq data in CODE (Consensus Optimization of Differentially Expressed gene detection).
}
\examples{
set.seed(123)
datarow=data_sample
group=group_sample
CODE.t_test(datarow,group)
}
\author{
Jiawei Zou
}
