% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/New_Metrics.R
\name{z_score}
\alias{z_score}
\title{Calculate Z score for data}
\usage{
z_score(data)
}
\arguments{
\item{data}{data, numberic verctor.}
}
\value{
Zscore for data.
}
\description{
Calculate Z score for data
}
