% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{group_sample}
\alias{group_sample}
\title{group information of dataset.}
\format{
A factor length 462
\describe{
\item{factor 1}{activated CD4+ T cells}
\item{factor 2}{naive CD4+ T cells}
...
}
}
\source{
\url{https://www.science.org/doi/full/10.1126/science.aah4115}
}
\usage{
group_sample
}
\description{
A vector of 1 and 2, the group information of the cells. Here, 1 for activated CD4+ T cells, 2 for naive CD4+ T cells.
}
\keyword{datasets}
